requirements.txt
fastapi
uvicorn
streamlit
pillow
requests
python-multipart
python-dotenv
anthropic
PyPDF2
httpx